Minor adjustments
Commented out the stuff from the first part of the assignment.

Made a little for loop as inspiration, which will need to be filled out more later.

Will still want to add a fragflag, after we get the segmentation bit down.

Trying to label everything I'm editting with a "Hugh" comment